+++ CHANGELOG.rst
@@ -16,8 +16,12 @@ This document contains the FSL-MRS release history in reverse chronological orde
 - Model methods in `fsl_mrs/models` now take a slice object argument rather than `first` and `last` values.
 - `io` utils have been updated to return a `Basis` object rather than `(FID, names, headers)`
 - `fsl_mrs_verify` target value has been updated to match the above changes.
-- Tests have been updated to match the above changes.
 - Added basis set read and conversion for `.mrui` formatted files.
+- Add complex cross-correlation frequency/phase alignment functionality, similar to that described in https://doi.org/10.1002/mrm.30252 (Deelchand, MRM, 2024).
+- `fsl_mrs_proc mrsi-align` now uses complex cross-correlation to perform phase and frequency alignment.
+- A target spectrum for alignment can be created using the new `basis2spectrum` function.
+- `fsl_mrs_proc align` now has the option to use complex cross-correlation `--method xcorr` to perform phase and frequency alignment. It is recommended to use the `--reference` option to assign a target with this.
+- `fsl_mrs_proc align` now has a new default option for the `--dim` input, which is set to "all" (previously "DIM_DYN"). This eases the use fo this function for phase/frequency alignment of a single transient. 
 
 2.4.14 (Wednesday 27th May 2026)
 -----------------------------------